+++ src/test/java/io/jenkins/plugins/git/forensics/reference/GitCommitListenerITest.java
@@ -1,11 +1,10 @@
-package io.jenkins.plugins.git.forensics.blame;
+package io.jenkins.plugins.git.forensics.reference;
 
 import hudson.model.Action;
 import hudson.model.FreeStyleProject;
 import hudson.model.Result;
 import hudson.plugins.git.GitSCM;
 import hudson.scm.SCM;
-import io.jenkins.plugins.git.forensics.reference.GitCommit;
 import jenkins.plugins.git.GitSampleRepoRule;
 import org.junit.ClassRule;
 import org.junit.Rule;
@@ -13,13 +12,17 @@ import org.junit.Test;
 import org.jvnet.hudson.test.JenkinsRule;
 
 import java.io.IOException;
-import java.util.Optional;
 
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.assertFalse;
 
+/**
+ * Tests the class {@link GitCommit}.
+ *
+ * @author Arne Schöntag
+ */
 @SuppressWarnings("PMD.SignatureDeclareThrowsException")
-public class GitCommitITest {
+public class GitCommitListenerITest {
 
     /**
      * Jenkins rule per suite.
@@ -34,28 +37,23 @@ public class GitCommitITest {
     public GitSampleRepoRule gitRepo = new GitSampleRepoRule();
 
     @Test
-    public void shouldFindReferencePoint() throws Exception {
+    public void shouldLogAllGitCommits() throws Exception {
         gitRepo.init();
-        createAndCommitFile("Test.java", "public class Test {}");
+        createAndCommitFile("First.java", "first commit after init");
+        createAndCommitFile("Second.java", "second commit after init");
 
-        FreeStyleProject reference = createFreeStyleProject("reference", new GitSCM(gitRepo.toString()));
-        JENKINS_PER_SUITE.assertBuildStatus(Result.SUCCESS, reference.scheduleBuild2(0, new Action[0]));
-
-        GitCommit referenceGitCommit = reference.getLastCompletedBuild().getAction(GitCommit.class);
-        assertEquals(2, referenceGitCommit.getRevisions().size());
-
-        createAndCommitFile("Branch.java", "another branch");
-
-        GitSCM git = new GitSCM(gitRepo.toString());
-        FreeStyleProject job = createFreeStyleProject("job", git);
+        FreeStyleProject job = createFreeStyleProject("job", new GitSCM(gitRepo.toString()));
         JENKINS_PER_SUITE.assertBuildStatus(Result.SUCCESS, job.scheduleBuild2(0, new Action[0]));
 
-        GitCommit jobGitCommit = job.getLastCompletedBuild().getAction(GitCommit.class);
-        assertEquals(3, jobGitCommit.getRevisions().size());
+        GitCommit gitCommit = job.getLastCompletedBuild().getAction(GitCommit.class);
+        assertFalse(gitCommit.getRevisions().isEmpty());
+        assertEquals(3, gitCommit.getRevisions().size());
 
-        Optional<String> result = jobGitCommit.getReferencePoint(referenceGitCommit, 100);
-        assertTrue(result.isPresent());
-        assertEquals(reference.getLastCompletedBuild().getExternalizableId(), result.get());
+        createAndCommitFile("Third.java", "third commit after init");
+        JENKINS_PER_SUITE.assertBuildStatus(Result.SUCCESS, job.scheduleBuild2(0, new Action[0]));
+
+        gitCommit = job.getLastCompletedBuild().getAction(GitCommit.class);
+        assertEquals(1, gitCommit.getRevisions().size());
     }
 
     private void createAndCommitFile(final String fileName, final String content) throws Exception {
